+++ src/index.js
@@ -3,6 +3,7 @@ import L from "leaflet";
 import "leaflet.path.drag";
 import Utils from "./scripts/utils.js";
 import { Rectangle, Circle, Path, Cable, MarkerItems, Socket } from "./scripts/mapitems.js";
+import items from "./scripts/items.js";
 
 
 class ItemManager {
@@ -301,114 +302,6 @@ window.categoryLayers = {};
 var layerControl = L.control.layers(baseMaps, categoryLayers).addTo(map);
 
 window.objects = new ItemManager(map);
-
-let items = [
-	{
-		"name": "6x12m Zelt (HaDiKo)",
-		"width": 12,
-		"height": 6,
-		"color": "#ffffff",
-		"category": "Zelte",
-	},
-	{
-		"name": "5x10m Zelt (K2-Bar)",
-		"width": 10,
-		"height": 5,
-		"color": "#ffffff",
-		"category": "Zelte",
-	},
-	{
-		"name": "3x3m Pavillon",
-		"width": 3,
-		"height": 3,
-		"color": "#ffffff",
-		"category": "Zelte",
-	},
-	{
-		"name": "Doppel-Pavillon",
-		"width": 6,
-		"height": 3,
-		"color": "#ffffff",
-		"category": "Zelte",
-	},
-	{
-		"name": "Rothaus-Bierinsel",
-		"width": 7.5,
-		"height": 7.5,
-		"color": "#FFFF00",
-		"category": "Wagen",
-	},
-	{
-		"name": "Biertisch",
-		"width": 2,
-		"height": 0.5,
-		"color": "#ffa500",
-		"category": "Biergarnituren",
-	},
-	{
-		"name": "Biertisch überbreit",
-		"width": 2.2,
-		"height": 0.8,
-		"color": "#FAB57F",
-		"category": "Biergarnituren",
-	},
-	{
-		"name": "Bierbank",
-		"width": 2,
-		"height": 0.25,
-		"color": "#ffa500",
-		"category": "Biergarnituren",
-	},
-	{
-		"name": "Biertheke",
-		"width": 2,
-		"height": 0.5,
-		"color": "#ff0000",
-		"category": "Biergarnituren",
-	},
-	{
-		"name": "Kühlschrank",
-		"width": 1,
-		"height": 1,
-		"color": "#2f4f4f",
-		"category": "Kühlgeräte",
-	},
-	{
-		"name": "Eistruhe",
-		"width": 1.5,
-		"height": 1,
-		"color": "#2f2f2f",
-		"category": "Kühlgeräte",
-	},
-	{
-		"name": "Spüle",
-		"width": 1.5,
-		"height": 1,
-		"color": "#BCC6CC",
-		"category": "Sanitär",
-	},
-	{
-		"name": "Bauzaun",
-		"width": 3.5,
-		"height": 0.04,
-		"color": "#808080",
-		"category": "Abgrenzungen",
-	},
-	{
-		"name": "1x2m Tisch",
-		"width": 2.0,
-		"height": 1.0,
-		"color": "#0000FF",
-		"category": "Tische",
-	},
-	{
-		"name": "Sofa",
-		"width": 2.5,
-		"height": 1.0,
-		"color": "#a52a2a",
-		"category": "Sofas",
-	},
-];
 
 L.Control.ItemAddControl = L.Control.extend({
 	onAdd: function (map) {
